fixed deprecation in `GlowESP`

+++ src/main/kotlin/com/charlatano/scripts/esp/GlowESP.kt
@@ -87,7 +87,7 @@ private fun Entity.chams(color: Color) {
 		csgoEXE[this + 0x70] = color.red.toByte()
 		csgoEXE[this + 0x71] = color.green.toByte()
 		csgoEXE[this + 0x72] = color.blue.toByte()
-		csgoEXE[this + 0x73] = color.alpha.toByte()
+		csgoEXE[this + 0x73] = color.alpha.toInt().toByte()
 		
 		engineDLL[dwModelAmbientMin] =
 			CHAMS_BRIGHTNESS.toFloat().hashCode() xor (engineDLL.address + dwModelAmbientMin - 0x2C).toInt()
